direct copying and make sure basic prod/rec works
- still need to check that listening rules can work
