+++ python/fire/cfg/__init__.py
@@ -1,0 +1,13 @@
+"""cfg
+
+Basic python configuration for fire
+
+This init file imports the configuration classes that
+are supposed to be used by the end user.
+"""
+
+from ._process import Process
+from ._processor import Processor, Producer, Analyzer
+from ._conditions import ConditionsProvider
+from ._storage import DropKeepRule
+from ._output_file import OutputFile
